+++ .idea/misc.xml
@@ -21,6 +21,7 @@
         <entry key="../../../../../../layout/compose-model-1649894034866.xml" value="0.20462962962962963" />
         <entry key="../../../../../../layout/compose-model-1649894128430.xml" value="0.25" />
         <entry key="../../../../../../layout/compose-model-1649904234615.xml" value="0.7030075187969925" />
+        <entry key="../../../../../../layout/compose-model-1649906407164.xml" value="0.1" />
         <entry key="../../../../layout/compose-model-1649172693920.xml" value="0.6754385964912281" />
       </map>
     </option>
